Add part3 - including db storage and user handling
This commit added several updates and new features to our system of
microservices. First, docker-compose was introduced for easier
handling/coordination of the services. Second, maintenance and general
readability was improved by refactoring the service-code into
appropriate files (although not into different packages).
Persistant storage was introduced to the consignment-service,
using MongoDB for that purpose. A new service (user-service) was also
added, along with a cli-tool (user-cli) to test it. The user-service
uses Postgres for data storage.

This commit was quite substantial, and users who want more details
regarding the steps taken should checkout branch 'part3'.